Use 'kill -9' to kill seahub.

+++ scripts/seahub.sh
@@ -229,9 +229,12 @@ function clear_sessions () {
 
 function stop_seahub () {
     if [[ -f ${pidfile} ]]; then
-        pid=$(cat "${pidfile}")
         echo "Stopping seahub ..."
-        kill ${pid}
+        pkill -9 -f "thirdpart/gunicorn"
+        if pgrep -f "thirdpart/gunicorn" 2>/dev/null 1>&2 ; then
+            echo 'Failed to stop seahub.'
+            exit 1
+        fi
         rm -f ${pidfile}
         return 0
     else
